requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
alpaca-trade-api==3.0.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.14.0
keras==2.14.0
yfinance==0.2.31
polygon-api-client==1.12.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
mlflow==2.8.1
optuna==3.4.0
pytest==7.4.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 